[log] Support level in each entry
- no longer use `entry.Logger.Level` to check if we should log, use
`entry.EntryLevel`
- `SetLevel` on logger would update `entry.EntryLevel` of all the
registered entries
- use bytes.buffer to test logging output with multiwriter
