+++ log/entry_generated.go.tmpl
@@ -6,13 +6,13 @@ import (
 
 {{ range .}}
 func (entry *Entry) {{.}}(args ...interface{}) {
-	if entry.Logger.Level >= {{.}}Level {
+	if entry.EntryLevel >= {{.}}Level {
 		entry.log({{.}}Level, fmt.Sprint(args...))
 	}
 }
 
 func (entry *Entry) {{.}}f(format string, args ...interface{}) {
-	if entry.Logger.Level >= {{.}}Level {
+	if entry.EntryLevel >= {{.}}Level {
 		entry.log({{.}}Level, fmt.Sprintf(format, args...))
 	}
 }
