+++ log/logger.go
@@ -1,132 +1,15 @@
 package log
 
 import (
+	"fmt"
 	"io"
 	"os"
 	"runtime"
 	"strings"
 	"sync"
 
-	"fmt"
 	"github.com/pkg/errors"
 )
-
-// Level is log level
-type Level uint8
-
-const (
-	// FatalLevel log error and call `os.Exit(1)`
-	FatalLevel Level = iota
-	// PanicLevel log error and call `panic`
-	PanicLevel
-	// ErrorLevel log error
-	ErrorLevel
-	// WarnLevel log warning
-	WarnLevel
-	// InfoLevel log info
-	InfoLevel
-	// DebugLevel log debug message, user should enable DebugLevel logging when report bug
-	DebugLevel
-	// TraceLevel is very verbose, user should enable it only on packages they are currently investing instead of globally
-	TraceLevel
-)
-
-// ShortUpperString returns the first 4 characters of a level in upper case
-func (level Level) ShortUpperString() string {
-	switch level {
-	case FatalLevel:
-		return "FATA"
-	case PanicLevel:
-		return "PANI"
-	case ErrorLevel:
-		return "ERRO"
-	case WarnLevel:
-		return "WARN"
-	case InfoLevel:
-		return "INFO"
-	case DebugLevel:
-		return "DEBU"
-	case TraceLevel:
-		return "TRAC"
-	default:
-		return "UNKN"
-	}
-}
-
-func (level Level) String() string {
-	switch level {
-	case FatalLevel:
-		return "fatal"
-	case PanicLevel:
-		return "panic"
-	case ErrorLevel:
-		return "error"
-	case WarnLevel:
-		return "warn"
-	case InfoLevel:
-		return "info"
-	case DebugLevel:
-		return "debug"
-	case TraceLevel:
-		return "trace"
-	default:
-		return "unknown"
-	}
-}
-
-// ParseLevel match the level string with Level, it will use strings.HasPrefix in non strict mode
-func ParseLevel(s string, strict bool) (Level, error) {
-	str := strings.ToLower(s)
-	if strict {
-		switch str {
-		case "fatal":
-			return FatalLevel, nil
-		case "panic":
-			return PanicLevel, nil
-		case "error":
-			return ErrorLevel, nil
-		case "warn":
-			return WarnLevel, nil
-		case "info":
-			return InfoLevel, nil
-		case "debug":
-			return DebugLevel, nil
-		case "trace":
-			return TraceLevel, nil
-		default:
-			return Level(250), errors.Errorf("unknown log level %s", str)
-		}
-	}
-	switch {
-	case strings.HasPrefix(str, "f"):
-		return FatalLevel, nil
-	case strings.HasPrefix(str, "p"):
-		return PanicLevel, nil
-	case strings.HasPrefix(str, "e"):
-		return ErrorLevel, nil
-	case strings.HasPrefix(str, "w"):
-		return WarnLevel, nil
-	case strings.HasPrefix(str, "i"):
-		return InfoLevel, nil
-	case strings.HasPrefix(str, "d"):
-		return DebugLevel, nil
-	case strings.HasPrefix(str, "t"):
-		return TraceLevel, nil
-	default:
-		return Level(250), errors.Errorf("unknown log level %s", str)
-	}
-}
-
-// AllLevels includes all the logging level
-var AllLevels = []Level{
-	FatalLevel,
-	PanicLevel,
-	ErrorLevel,
-	WarnLevel,
-	InfoLevel,
-	DebugLevel,
-	TraceLevel,
-}
 
 // Fields is key-value string pair to annotate the log and can be used by filter
 type Fields map[string]string
@@ -159,6 +42,25 @@ func NewLogger() *Logger {
 	return l
 }
 
+func (log *Logger) SetLevel(s string) error {
+	newLevel, err := ParseLevel(s, false)
+	if err != nil {
+		return errors.WithMessage(err, fmt.Sprintf("can't set logging level to %s", s))
+	}
+	if log.Level == newLevel {
+		return nil
+	}
+	// update all the registered entries
+	for _, entry := range log.Entries {
+		// only update entry's level if the new global level is more verbose
+		if newLevel > entry.EntryLevel {
+			entry.EntryLevel = newLevel
+		}
+	}
+	log.Level = newLevel
+	return nil
+}
+
 func (log *Logger) ApplyConfig(c *Config) error {
 	if err := c.Validate(); err != nil {
 		return err
@@ -166,7 +68,7 @@ func (log *Logger) ApplyConfig(c *Config) error {
 	if log.Level.String() != c.Level {
 		newLevel, err := ParseLevel(c.Level, false)
 		if err != nil {
-			return errors.Wrapf(err, "can't set logging level to %s", c.Level)
+			return errors.WithMessage(err, fmt.Sprintf("can't set logging level to %s", c.Level))
 		}
 		log.Level = newLevel
 	}
@@ -194,11 +96,14 @@ func (log *Logger) AddFilter(filter Filter, level Level) {
 }
 
 // NewEntry returns an Entry with empty Fields
+// Deprecated: use RegisterPkg instead
 func (log *Logger) NewEntry() *Entry {
 	// TODO: may use pool, but need benchmark to see if using pool provides improvement
 	return &Entry{
-		Logger: log,
-		Fields: make(map[string]string, 1),
+		Logger:     log,
+		Pkg:        "",
+		EntryLevel: log.Level,
+		Fields:     make(map[string]string, 1),
 	}
 }
 
@@ -208,9 +113,10 @@ func (log *Logger) NewEntryWithPkg(pkgName string) *Entry {
 	fields := make(map[string]string, 1)
 	fields["pkg"] = pkgName
 	e := &Entry{
-		Logger: log,
-		Fields: fields,
-		Pkg:    pkgName,
+		Logger:     log,
+		Pkg:        pkgName,
+		EntryLevel: log.Level,
+		Fields:     fields,
 	}
 	log.Entries[pkgName] = e
 	return e
@@ -223,9 +129,10 @@ func (log *Logger) RegisterPkg() *Entry {
 	pkg := getCallerPackage(2)
 	fields["pkg"] = pkg
 	e := &Entry{
-		Logger: log,
-		Fields: fields,
-		Pkg:    pkg,
+		Logger:     log,
+		Pkg:        pkg,
+		EntryLevel: log.Level,
+		Fields:     fields,
 	}
 	log.Entries[pkg] = e
 	return e
